FAR-19/refactor: 게시글 리팩토링
