+++ src/main/java/com/ssafy/faraway/domain/post/repository/PostQueryRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 
 public interface PostQueryRepository {
     Post searchById(Long postId);
-    List<ListPostResponse> searchByCondition(PostSearchCondition searchCondition, Pageable pageable);
+    List<ListPostResponse> searchByCondition(PostSearchCondition condition, Pageable pageable);
 }
